pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swiggy-analyzer"
version = "0.1.0"
description = "Swiggy Instamart buying pattern analyzer and recommendation engine"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Suvrat Hiran", email = "suvrat.hiran@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "requests>=2.31.0",
    "httpx>=0.24.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
    "keyring>=24.0.0",
    "cryptography>=41.0.0",
    "authlib>=1.2.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
swiggy-analyzer = "swiggy_analyzer.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["swiggy_analyzer*"]
